[imodel-content-tree] Update to 3.0 (#169)
