+++ README.md
@@ -20,11 +20,10 @@ Each one has it's own folder and README describing the package.
 
 ### iTwin.js@^2 Compatible
 
-| Name                               | Folder                                                        | Changelog                                                        | Version                                                                                                                                                |
-| ---------------------------------- | ------------------------------------------------------------- | ---------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------ |
-| @bentley/imodel-content-tree-react | [imodel-content-tree](./packages/bentley/imodel-content-tree) | [changelog](./packages/bentley/imodel-content-tree/CHANGELOG.md) | [![npm version](https://badge.fury.io/js/%40bentley%2Fimodel-content-tree-react.svg)](https://badge.fury.io/js/%40bentley%2Fimodel-content-tree-react) |
-| @bentley/imodel-react-hooks        | [imodel-react-hooks](./packages/bentley/imodel-react-hooks)   | [changelog](./packages/bentley/imodel-react-hooks/CHANGELOG.md)  | [![npm version](https://badge.fury.io/js/%40bentley%2Fimodel-react-hooks.svg)](https://badge.fury.io/js/%40bentley%2Fimodel-react-hooks)               |
-| @bentley/geo-tools-react           | [geo-tools](./packages/bentley/geo-tools)                     | [changelog](./packages/bentley/geo-tools/CHANGELOG.md)           | [![npm version](https://badge.fury.io/js/%40bentley%2Fgeo-tools-react.svg)](https://badge.fury.io/js/%40bentley%2Fgeo-tools-react)                     |
+| Name                        | Folder                                                      | Changelog                                                       | Version                                                                                                                                  |
+| --------------------------- | ----------------------------------------------------------- | --------------------------------------------------------------- | ---------------------------------------------------------------------------------------------------------------------------------------- |
+| @bentley/imodel-react-hooks | [imodel-react-hooks](./packages/bentley/imodel-react-hooks) | [changelog](./packages/bentley/imodel-react-hooks/CHANGELOG.md) | [![npm version](https://badge.fury.io/js/%40bentley%2Fimodel-react-hooks.svg)](https://badge.fury.io/js/%40bentley%2Fimodel-react-hooks) |
+| @bentley/geo-tools-react    | [geo-tools](./packages/bentley/geo-tools)                   | [changelog](./packages/bentley/geo-tools/CHANGELOG.md)          | [![npm version](https://badge.fury.io/js/%40bentley%2Fgeo-tools-react.svg)](https://badge.fury.io/js/%40bentley%2Fgeo-tools-react)       |
 
 ### Archived
