+++ common/changes/@itwin/geo-tools-react/imodel-content-tree-update-to-3.0_2022-02-28-23-33.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/geo-tools-react",
+      "comment": "",
+      "type": "none"
+    }
+  ],
+  "packageName": "@itwin/geo-tools-react"
+}
