+++ common/changes/@itwin/tree-widget-react/imodel-content-tree-update-to-3.0_2022-02-28-23-33.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/tree-widget-react",
+      "comment": "Add `IModelContentTree`",
+      "type": "minor"
+    }
+  ],
+  "packageName": "@itwin/tree-widget-react"
+}
