+++ common/config/rush/version-policies.json
@@ -20,11 +20,6 @@
     "lockedMajor": 0
   },
   {
-    "policyName": "imodel-content-tree",
-    "definitionName": "individualVersion",
-    "lockedMajor": 0
-  },
-  {
     "policyName": "imodel-react-hooks",
     "definitionName": "individualVersion",
     "lockedMajor": 0
