--- packages/bentley/imodel-content-tree/.npmignore
@@ -1,9 +1,0 @@
-# start off ignoring everything, and then add back only the files we want
-*
-!lib/**/*.d.ts
-!lib/**/*.d.ts.map
-!lib/**/*.js
-!lib/**/*.js.map
-!lib/**/*.json
-!lib/**/*.*css
-!lib/public/**/*
