--- packages/bentley/imodel-content-tree/CHANGELOG.json
@@ -1,124 +1,0 @@
-{
-  "name": "@bentley/imodel-content-tree-react",
-  "entries": [
-    {
-      "version": "0.3.3",
-      "tag": "@bentley/imodel-content-tree-react_v0.3.3",
-      "date": "Tue, 23 Nov 2021 21:19:42 GMT",
-      "comments": {
-        "patch": [
-          {
-            "comment": "Load InformationContentElement nodes under geometric models"
-          }
-        ]
-      }
-    },
-    {
-      "version": "0.3.2",
-      "tag": "@bentley/imodel-content-tree-react_v0.3.2",
-      "date": "Tue, 08 Jun 2021 21:23:59 GMT",
-      "comments": {
-        "patch": [
-          {
-            "comment": "bump classnames dep"
-          }
-        ]
-      }
-    },
-    {
-      "version": "0.3.1",
-      "tag": "@bentley/imodel-content-tree-react_v0.3.1",
-      "date": "Thu, 13 May 2021 21:15:14 GMT",
-      "comments": {
-        "patch": [
-          {
-            "comment": "induce version bump"
-          }
-        ]
-      }
-    },
-    {
-      "version": "0.3.0",
-      "tag": "@bentley/imodel-content-tree-react_v0.3.0",
-      "date": "Thu, 13 May 2021 16:38:52 GMT",
-      "comments": {
-        "minor": [
-          {
-            "comment": "update to imjs 2.15.2 to resolve breaking change to usePresentationTreeNodeLoader"
-          }
-        ]
-      }
-    },
-    {
-      "version": "0.2.0",
-      "tag": "@bentley/imodel-content-tree-react_v0.2.0",
-      "date": "Fri, 26 Feb 2021 18:43:33 GMT",
-      "comments": {
-        "minor": [
-          {
-            "comment": "Directly under non-geometric models node show only elements that have no parent"
-          },
-          {
-            "comment": "Group elements under non-geometric models by class"
-          },
-          {
-            "comment": "For every element node under non-geometric model load its child elements as child nodes"
-          }
-        ]
-      }
-    },
-    {
-      "version": "0.1.0",
-      "tag": "@bentley/imodel-content-tree-react_v0.1.0",
-      "date": "Fri, 19 Feb 2021 15:47:25 GMT",
-      "comments": {
-        "patch": [
-          {
-            "comment": "Load member elements for GroupInformationElement nodes"
-          }
-        ],
-        "minor": [
-          {
-            "comment": "Update imodeljs dependencies to ^2.7"
-          }
-        ]
-      }
-    },
-    {
-      "version": "0.0.4",
-      "tag": "@bentley/imodel-content-tree-react_v0.0.4",
-      "date": "Fri, 16 Oct 2020 19:19:15 GMT",
-      "comments": {
-        "patch": [
-          {
-            "comment": "Handle GraphicalPartition3d similar to PhysicalPartition - it should not be displayed if there's a 'GraphicalPartition3d.Model.Content' attribute in JsonProperties"
-          }
-        ]
-      }
-    },
-    {
-      "version": "0.0.3",
-      "tag": "@bentley/imodel-content-tree-react_v0.0.3",
-      "date": "Wed, 07 Oct 2020 18:30:54 GMT",
-      "comments": {
-        "patch": [
-          {
-            "comment": "Fix 'main' and 'typings' fields by removing the '-react' to match the name of the barrel file instead of package name."
-          }
-        ]
-      }
-    },
-    {
-      "version": "0.0.2",
-      "tag": "@bentley/imodel-content-tree-react_v0.0.2",
-      "date": "Wed, 26 Aug 2020 12:14:46 GMT",
-      "comments": {
-        "patch": [
-          {
-            "comment": "Initial commit"
-          }
-        ]
-      }
-    }
-  ]
-}
