--- packages/bentley/imodel-content-tree/src/imodel-content-tree.ts
@@ -1,5 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-export * from "./components/IModelContentTree";
