--- packages/bentley/imodel-content-tree/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "./node_modules/@bentley/build-tools/tsconfig-base.json",
-  "compilerOptions": {
-    "target": "es2017",
-    "module": "commonjs",
-    "declarationMap": true,
-    "outDir": "./lib",
-    "skipLibCheck": true,
-    "esModuleInterop": true,
-    "jsx": "react",
-    "preserveSymlinks": true,
-    "resolveJsonModule": true
-  },
-  "include": ["./src/**/*.ts*"],
-  "exclude": ["lib", "node_modules"]
-}
