+++ packages/itwin/geo-tools/src/GeoToolsUiItemsProvider.tsx
@@ -1,3 +1,7 @@
+/*---------------------------------------------------------------------------------------------
+* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+* See LICENSE.md in the project root for license terms and full copyright notice.
+*--------------------------------------------------------------------------------------------*/
 import type {
   CommonToolbarItem,
   UiItemsProvider,
