--- packages/itwin/tree-widget/src/components/rulesets/StoriesBreakdown.json
@@ -1,61 +1,0 @@
-{
-  "$schema": "../../node_modules/@itwin/presentation-common/Ruleset.schema.json",
-  "id": "tree-widget-react/StoriesBreakdown",
-  "rules": [
-    {
-      "ruleType": "RootNodes",
-      "specifications": [
-        {
-          "specType": "InstanceNodesOfSpecificClasses",
-          "classes": [
-            {
-              "schemaName": "BuildingSpatial",
-              "classNames": ["Story"]
-            }
-          ],
-          "arePolymorphic": true,
-          "relatedInstances": [
-            {
-              "relationshipPath": {
-                "relationship": {
-                  "schemaName": "BisCore",
-                  "className": "ModelContainsElements"
-                },
-                "direction": "Backward"
-              },
-              "isRequired": true,
-              "alias": "model"
-            }
-          ],
-          "instanceFilter": "NOT model.IsPrivate AND this.HasRelatedInstance(\"SpatialComposition:CompositeComposesSubComposites\", \"Forward\", \"BuildingSpatial:Space\")",
-          "groupByClass": false,
-          "groupByLabel": false
-        }
-      ]
-    },
-    {
-      "ruleType": "ChildNodes",
-      "condition": "ParentNode.IsOfClass(\"Story\", \"BuildingSpatial\")",
-      "specifications": [
-        {
-          "specType": "RelatedInstanceNodes",
-          "relationshipPaths": [
-            {
-              "relationship": {
-                "schemaName": "SpatialComposition",
-                "className": "CompositeComposesSubComposites"
-              },
-              "direction": "Forward",
-              "targetClass": {
-                "schemaName": "BuildingSpatial",
-                "className": "Space"
-              }
-            }
-          ],
-          "groupByClass": false,
-          "groupByLabel": false
-        }
-      ]
-    }
-  ]
-}
