+++ packages/itwin/tree-widget/src/components/trees/index.ts
@@ -2,8 +2,10 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
+
 export * from "./ModelsTree";
 export * from "./CategoriesTree";
 export * from "./SpatialTree";
 export * from "./ClassificationsTree";
 export * from "./TreeWithRuleset";
+export * from "./IModelContentTree";
